Updated tables primary keys type to bigInteger

+++ database/migrations/2018_12_16_000420_create_media_table.php
@@ -14,7 +14,7 @@ class CreateMediaTable extends Migration
     public function up()
     {
         Schema::create('media', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('name')->index()->nullable();
             $table->string('relative_path');
             $table->integer('width')->nullable();
@@ -25,13 +25,13 @@ class CreateMediaTable extends Migration
             $table->unsignedInteger('mediable_id')->nullable();
             $table->string('mediable_type')->nullable();
 
-            $table->unsignedInteger('media_collection_id')->nullable();
+            $table->unsignedBigInteger('media_collection_id')->nullable();
             $table
                 ->foreign('media_collection_id')
                 ->references('id')
                 ->on('media_collections');
 
-            $table->unsignedInteger('user_id')->nullable();
+            $table->unsignedBigInteger('user_id')->nullable();
 
             $table->timestamps();
         });
